Changed Rect to floats. Added Color Type. Refactored SDL_renderer a bit.

+++ pong/include/engine/graphics/Rect.h
@@ -13,14 +13,13 @@ namespace graphics {
 
 struct Rect {
     Rect() : height{0}, width{0}, x{0}, y{0} {}
-    Rect(float width, float height) : height{height}, width{width}, x{0}, y{0} {}
-    Rect(float x, float y, float width, float height) : height{height}, width{width}, x{x}, y{y} {}
+    Rect(int width, int height) : height{height}, width{width}, x{0}, y{0} {}
+    Rect(int x, int y, int width, int height) : height{height}, width{width}, x{x}, y{y} {}
 
-    float height;
-    float width;
-    float x;
-    float y;
-
+    int height;
+    int width;
+    int x;
+    int y;
 };
 }
 }
